fix(components): dangerous errors needs client rendering
